django/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-simple-history==3.10.1
django-widget-tweaks==1.5.0
Faker==37.4.0
pillow==11.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
